requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
pillow
streamlit
matplotlib
transformers
torch
accelerate
streamlit-cropper
